rvfi-dii: Add rvfi-dii server that handles injected instructions
The servers imitates the bitbang server and is initialised with
a port number that waits for RVFI-DII commands and handles them
over a socket, then send the trace back.

rvfi-dii.h: Is a generic interface that is compatible with RVFI-DII
specifications for input and output packets.

rvfi-dii.cc: Is the implementation of the rvfi-dii server in Spike